+++ src/lib/core/directshow/dsfAnxMux/AnxPageInterleaver.cpp
@@ -76,7 +76,7 @@ void AnxPageInterleaver::addAnnodex_2_0_BOS() {
 
 void AnxPageInterleaver::addAllAnxData_2_0_BOS() {
 	OggPage* locOggPage = NULL;
-	for (int i = 0; i < mInputStreams.size() - 1; i++) {
+	for (size_t i = 0; i < mInputStreams.size() - 1; i++) {
 		locOggPage = mInputStreams[i]->popFront();
 		mBytesWritten += locOggPage->pageSize();
 		mFileWriter->acceptOggPage(locOggPage);
@@ -103,7 +103,7 @@ bool AnxPageInterleaver::gotAllHeaders() {
 
 	bool locWasAny = false;
 	bool locIsOK = true;
-	for (int i = 0; i < mInputStreams.size(); i++) {
+	for (size_t i = 0; i < mInputStreams.size(); i++) {
 		if (mInputStreams[i]->isActive()) {
 			
 			//if ((mInputStreams[i]->peekFront() != NULL) || (!mInputStreams[i]->isActive())) {
